Added templates for new user routes (not implemented yet) and made it so that only the creator of a group can modify/delete it + bug fixes

+++ backend/controllers/group_controller.js
@@ -15,10 +15,9 @@ const get_filtered_groups = async(req, res) => {
     res.status(200).json(groups)
 }
 
-/*
 // get a single group
 const get_group = async(req, res) => {
-    const {id} = req.params
+    const {id} = req.params.id
     
     if (!mongoose.Types.ObjectId.isValid(id)) {
         return res.status(404).json({error: "No such group"})
@@ -32,7 +31,6 @@ const get_group = async(req, res) => {
 
     res.status(200).json(group)
 }
-*/
 
 // create new group
 const create_group = async(req, res) => {
@@ -71,16 +69,17 @@ const create_group = async(req, res) => {
 
 // delete a group
 const delete_group = async(req, res) => {
-    const {id} = req.params
+    const id = req.params.id
+    const creator_id = req.user._id
 
     if (!mongoose.Types.ObjectId.isValid(id)) {
         return res.status(404).json({error: "No such group"})
     }
 
-    const group = await Group.findOneAndDelete({_id: id})
+    const group = await Group.findOneAndDelete({_id: id, creator_id})
 
     if (!group) {
-        return res.status(404).json({error: "No such group"})
+        return res.status(404).json({error: "Not the user who created the group"})
     }
 
     res.status(200).json(group)
@@ -88,8 +87,8 @@ const delete_group = async(req, res) => {
 
 // join a group
 const join_group = async(req, res) => {
-    const {id} = req.params
-    const {member_id} = req.user._id
+    const id = req.params.id
+    const member_id = req.user._id
 
     if (!mongoose.Types.ObjectId.isValid(id)) {
         return res.status(404).json({error: "No such group"})
@@ -116,25 +115,19 @@ const join_group = async(req, res) => {
 
 // update a workout
 const update_group = async(req, res) => {
-    const {id} = req.params
-    //const {creator_id} = req.user._id
+    const id = req.params.id
+    const creator_id = req.user._id
 
     if (!mongoose.Types.ObjectId.isValid(id)) {
         return res.status(404).json({error: "No such group"})
     }
 
-    const group = await Group.findOneAndUpdate({_id: id}, {
+    const group = await Group.findOneAndUpdate({_id: id, creator_id}, {
         ...req.body
     })
 
-    /*
-    if (group.creator_id !== creator_id) {
-        return res.status(401).json({error: "Not the user who created the group"})
-    }
-    */
-
     if (!group) {
-        return res.status(404).json({error: "No such group"})
+        return res.status(404).json({error: "Not the user who created the group"})
     }
 
     res.status(200).json(group)
@@ -143,7 +136,7 @@ const update_group = async(req, res) => {
 module.exports = {
     get_all_groups,
     get_filtered_groups,
-    //get_workout,
+    get_group,
     create_group,
     delete_group,
     update_group,
